Temp add norole
norole is a onetime function that runs user ids pulled from Carlbot and roles them into a specified role.

Caveat: This function is uses discord.js member cache and may not get all of the users.  This is a temp solution to lessen the number of user without any roles.